+++ AmazonChimeSDK/AmazonChimeSDK/internal/contentshare/DefaultContentShareVideoClientController.swift
@@ -100,6 +100,7 @@ import Foundation
 
     private func stopVideoClient() {
         videoClient.stop()
+        videoClient.delegate = nil
     }
 
     public func subscribeToVideoClientStateChange(observer: ContentShareObserver) {
@@ -150,7 +151,6 @@ extension DefaultContentShareVideoClientController: VideoClientDelegate {
             observer.contentShareDidStop(status: ContentShareStatus(statusCode: .videoServiceFailed))
         }
         isSharing = false
-        cleanUp()
     }
 
     public func videoClientDidStop(_ client: VideoClient?) {
@@ -160,7 +160,6 @@ extension DefaultContentShareVideoClientController: VideoClientDelegate {
             observer.contentShareDidStop(status: ContentShareStatus(statusCode: .ok))
         }
         isSharing = false
-        cleanUp()
     }
 
     public func videoClientMetricsReceived(_ metrics: [AnyHashable: Any]?) {
@@ -170,9 +169,5 @@ extension DefaultContentShareVideoClientController: VideoClientDelegate {
 
     private func resetContentShareVideoClientMetrics() {
         clientMetricsCollector.processContentShareVideoClientMetrics(metrics: [:])
-    }
-    
-    private func cleanUp() {
-        videoClient.delegate = nil
     }
 }