+++ AmazonChimeSDK/AmazonChimeSDK/internal/video/VideoClientProtocol.swift
@@ -11,6 +11,7 @@ import AmazonChimeSDKMedia
 import Foundation
 
 @objc public protocol VideoClientProtocol {
+    // TODO: look into make delegate weak preference
     var delegate: VideoClientDelegate! { get set }
 
     static func globalInitialize()