+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/MeetingModel.swift
@@ -606,7 +606,7 @@ extension MeetingModel: MetricsObserver {
             return
         }
         metricsModel.updateAppMetrics(metrics: metrics)
-        logger.info(msg: "Media metrics have been received: \(observableMetrics)")
+        logger.debug(debugFunction: { return "Media metrics have been received: \(observableMetrics)"})
         if activeMode == .metrics {
             metricsModel.metricsUpdatedHandler?()
         }